Move delete cname script to new folder (#2355)
